feat: added support for encrypting events persisted in sqlite3 db
