+++ packages/plugins/rudder_plugin_android/android/build.gradle
@@ -38,6 +38,6 @@ android {
 }
 
 dependencies {
-    implementation 'com.rudderstack.android.sdk:core:[1.17.0, 2.0.0)'
+    implementation 'com.rudderstack.android.sdk:core:[1.18.0, 2.0.0)'
     implementation 'com.google.code.gson:gson:2.8.6'
 }
