+++ packages/plugins/rudder_plugin_interface/lib/platform.dart
@@ -4,6 +4,8 @@ export 'src/constants.dart';
 export 'src/enums.dart';
 export 'src/rudder_logger.dart';
 export 'src/models/rudder_config.dart';
+export 'src/models/rudder_web_config.dart';
+export 'src/models/rudder_mobile_config.dart';
 export 'src/models/rudder_integration.dart';
 export 'src/models/rudder_option.dart';
 export 'src/models/rudder_property.dart';
